Moved FUSE device under the fuse directory
